restaurant isclosed提取 && menu modal
