+++ src/features/Menu/Menu.js
@@ -1,7 +1,6 @@
 import React,{ useEffect } from 'react';
 import { useDispatch,useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
-import moment from 'moment-timezone';
 
 import _ from 'lodash';
 
@@ -15,7 +14,7 @@ import { get,init } from '../../Common/Intl';
 import Header from '../../Components/Header/Header';
 import Cart from '../../Components/Cart/Cart';
 import MenuBox from '../../Components/MenuBox/MenuBox';
-import { showErrorAsync } from '../../Redux/Reducer/error';
+import { showErrorAsync ,hideError } from '../../Redux/Reducer/error';
 
 /* style */
 import './menu.scss';
@@ -23,6 +22,9 @@ import './menu.scss';
 /* action */
 import { sendRequestMenu,renderMenu } from './state/reducer';
 import { language } from '../../Redux/Reducer/header';
+
+/* function */
+import { isClosed } from '../../Common/utils';
 
 export default function Menu () {
 
@@ -55,38 +57,23 @@ export default function Menu () {
 
   /* 每30s检测 */
   useEffect(()=>{
+    checkOpen();
     let timer = setInterval(()=>{
-      let flag = checkOpen();
-      if(!flag){
-        dispatch(showErrorAsync('该店已关闭'));
-      }
-    },30000);
+      checkOpen();
+    },3000);
     return ()=>{
       if(timer){
         clearInterval(timer);
+        dispatch(hideError());
       }
     };
   },[]);
   /* 检测开关店 */
   function checkOpen (){
-    let date = new Date();
-    var newYork = moment.tz(date,'America/New_York');
-    let checkTime = newYork.hours() * 60 + newYork.minutes();
-    let isOpen = true;
-    /* 是否人为关闭 */
-    if(restInfo.closed){
-      isOpen = false;
-    }else{
-      /* 是否在开店时间内 */
-      _.map(restInfo.hours,(item)=>{
-        if(item.dayOfWeek == newYork.day() && item.start <= checkTime && checkTime <= item.end){
-          isOpen = true;
-        }else{
-          isOpen = false;
-        }
-      });
+    let flag = isClosed(restInfo);
+    if(flag){
+      dispatch(showErrorAsync('该店已关闭'));
     }
-    return isOpen;
   }
 
   /* 渲染菜单 */
